+++ test/Swashbuckle.AspNetCore.Filters.Test/Examples/ExamplesOperationFilterTests.cs
@@ -278,27 +278,5 @@ namespace Swashbuckle.AspNetCore.Filters.Test.Examples
             formatedExample.EndsWith('"').ShouldBeTrue();
             formatedExample.StartsWith('"').ShouldBeTrue();
         }
-
-        [Fact]
-        public void ShouldNotEmitObsoleteProperties()
-        {
-            // Arrange
-            schemaGeneratorOptions.IgnoreObsoleteProperties = true;
-            var response = new OpenApiResponse { Content = new Dictionary<string, OpenApiMediaType> { { "application/json", new OpenApiMediaType() } } };
-            var operation = new OpenApiOperation { OperationId = "foobar", Responses = new OpenApiResponses() };
-            operation.Responses.Add("200", response);
-
-            var filterContext = FilterContextFor(typeof(FakeActions), nameof(FakeActions.AnnotatedWithSwaggerResponseExampleAttribute));
-            SetSwaggerResponses(operation, filterContext);
-
-            // Act
-            sut.Apply(operation, filterContext);
-
-            // Assert
-            string jsonExample = ((OpenApiRawString)response.Content["application/json"].Example).Value;
-            var expectedExample = new PersonResponseExample().GetExamples();
-            jsonExample.ShouldNotContain($"\"age\": {expectedExample.Age}", Case.Sensitive);
-            jsonExample.ShouldContain($"\"id\": {expectedExample.Id}", Case.Sensitive);
-        }
     }
 }