+++ CHANGELOG.md
@@ -4,17 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [7.0.0] - 2021-04-03
 ### Security
 - PR #176 update NuGet packages
 
 ### Changed
 - PR #178 dependency cleanup
-- PR #179 introduce Swashbuckle.AspNetCore.Filters.Abstractions package so that IExamplesProvider can be consumed without Swashbuckle dependency.
 
 ### Added
+- PR #179 introduce Swashbuckle.AspNetCore.Filters.Abstractions package so that IExamplesProvider can be consumed without Swashbuckle dependency.
 - Target .NET 5.0
 - Microsoft.SourceLink.GitHub NuGet reference
+
+### Removed
+- Issue #174 remove ExcludeObsoletePropertiesResolver and NewtonsoftJson dependency.
 
 ## [6.1.0] - 2021-02-15
 ### Added